update posting functions & fixing minor things
